# docs/swagger-v3.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API сервиса постов для платформы SwiftTalk",
        "title": "SwiftTalk Post Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/post": {
            "get": {
                "description": "Получить список всех постов",
                "tags": [
                    "posts"
                ],
                "summary": "Получить список постов",
                "responses": {
                    "200": {
                        "description": "Успешный ответ с списком постов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.GetPostsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создать новый пост с заголовком, содержанием и опциональными изображениями",
                "tags": [
                    "posts"
                ],
                "summary": "Создать новый пост",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "description": "Заголовок поста",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Содержание поста",
                                        "type": "string"
                                    },
                                    "images": {
                                        "description": "Изображения (множественная загрузка)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "title",
                                    "content"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Успешное создание поста",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.CreatePostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в параметрах запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/{id}": {
            "get": {
                "description": "Получить детальную информацию о посте по его идентификатору",
                "tags": [
                    "posts"
                ],
                "summary": "Получить пост по ID",
                "parameters": [
                    {
                        "description": "ID поста",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ с информацией о посте",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.GetPostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновить пост по его идентификатору",
                "tags": [
                    "posts"
                ],
                "summary": "Обновить пост по ID",
                "parameters": [
                    {
                        "description": "ID поста",
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "description": "Заголовок поста",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Содержание поста",
                                        "type": "string"
                                    },
                                    "images": {
                                        "description": "Изображения (множественная загрузка)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "deleted_images": {
                                        "description": "Удаленные изображения",
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "title",
                                    "content"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление поста",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.UpdatePostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в параметрах запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить пост по его идентификатору",
                "tags": [
                    "posts"
                ],
                "summary": "Удалить пост по ID",
                "parameters": [
                    {
                        "description": "ID поста",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление поста",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.DeletePostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/bff.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "//localhost:5001/post-service"
        }
    ],
    "components": {
        "schemas": {
            "bff.CreatePostResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "bff.DeletePostResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "bff.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
            "bff.GetPostResponse": {
                "type": "object",
                "properties": {
                    "post": {
                        "$ref": "#/components/schemas/bff.Post"
                    }
                }
            },
            "bff.GetPostsResponse": {
                "type": "object",
                "properties": {
                    "limit": {
                        "type": "integer"
                    },
                    "page": {
                        "type": "integer"
                    },
                    "posts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/bff.Post"
                        }
                    },
                    "total": {
                        "type": "integer"
                    }
                }
            },
            "bff.Image": {
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    }
                }
            },
            "bff.Post": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "string"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/bff.Image"
                        }
                    },
                    "title": {
                        "type": "string"
                    },
                    "user_uuid": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    }
                }
            },
            "bff.UpdatePostResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            }
        }
    }
}